requirements.txt
functions-framework

# Core dependencies
fastapi
uvicorn
pydantic
firebase-admin
python-dotenv

# Trademark similarity calculations
python-levenshtein
metaphone

# Google Cloud & Vertex AI dependencies
google-genai
google-cloud-core
google-api-core
google-cloud-aiplatform

# Client utilities
httpx

# Development dependencies
pytest
pytest-asyncio
ruff 